Restore SerializableException.ToString behavior
The original implementation of `SerializableException.ToString`
was serialising an exception information as XML, that restricted
usability of `SerializableException.ToString`  to relatively narrow
set of applications.

The change introduces a new method `ToXmlString` that returns
an XML-serialised exception information.
A call to `ToString` will return a string representation of the underlying
exception.

The implementation also lacked access to the original exception,
which is being fixed as part of this change.